+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react"
 import ReactDOM from "react-dom"
 
-import './style.css' 
+import './assets/styles/index.css' 
 import App from "./App"
 
 ReactDOM.render(
